Clean up the issues in refactored Timer
I'm still reading about what the best practices are for using
namespaces, but I'm not totally on board with the idea of sticking all
the timing stuff in G::time::. Originally it was just going in time::
but time was taken by the time function in <time.h> so I stuck it under
the "globals/game" namespace "G". I've also contemplated doing
old-fashioned namespacing by prefixing my constant globals with G_ and
getting rid of G:: entirely. Regardless, I'm still unsure of the
standard practice for namespace use in C++.

Also, I fixed all of the warnings and added pragma directives to ignore
3rd party warnings. (I need to check if those compile on GCC, or other
Linux-brand compilers)

I'm also slowly beginning to reformat the code to my now preferred
style.

I'm still thinking about the structure I want to refactor this into, but
I think I have a pretty good idea. I have to do a few more small edits
with the repo before doing intensive refactoring of major components.

+++ Sources/debug.cpp
@@ -2,7 +2,7 @@
 
 D(
 void slowPrint(int middle, int after, std::string txt) {
-	for (int i = 0; i < txt.size() - 1; i++) {
+	for (int i = 0; i < static_cast<int>(txt.size() - 1); i++) {
 		std::cout << txt[i];
 		std::this_thread::sleep_for(std::chrono::milliseconds(middle));
 	}
